requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
astroid==2.8.6
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.0.12
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
loguru==0.5.3
MarkupSafe==2.1.3
mccabe==0.6.1
packaging==23.2
pageobject==0.0.52
pi==0.1.2
platformdirs==3.11.0
pluggy==1.3.0
py==1.11.0
pylint==2.11.1
pytest~=7.4.2
pytest-base-url==2.0.0
pytest-html==4.0.2
pytest-metadata==3.0.0
pytest-selenium==2.0.1
pytest-variables==3.0.0
requests==2.26.0
selenium~=4.14.0
six==1.16.0
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.18
wrapt==1.13.3
